requirements.txt
streamlit
pandas
altair
google-generativeai
gspread
oauth2client
